requirements.txt
fastapi
uvicorn
python-multipart
pytest
httpx
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv